+++ src/components/Badge/index.tsx
@@ -3,7 +3,7 @@ import styled from "@emotion/styled";
 export interface BadgeProps { }
 
 export function Badge({ }: BadgeProps) {
-  return <StyledBadge data-testid="badge" >A</StyledBadge>;
+  return <StyledBadge data-testid="badge" />
 }
 
 const StyledBadge = styled.span`