+++ src/components/Button/index.tsx
@@ -1,13 +1,17 @@
 import styled from "@emotion/styled";
 
-export interface ButtonProps {
+export interface ButtonProps extends ButtonBaseProps {
   onClick: () => void;
   label: string;
 }
 
-export function Button({ onClick, label }: ButtonProps) {
+interface ButtonBaseProps {
+  block?: boolean;
+}
+
+export function Button({ onClick, label, block = false }: ButtonProps) {
   return (
-    <StyledButton data-testid="button" onClick={onClick}>
+    <StyledButton data-testid="button" onClick={onClick} block={block}>
       {label}
     </StyledButton>
   );
@@ -15,10 +19,11 @@ export function Button({ onClick, label }: ButtonProps) {
 
 const StyledButton = styled.button`
   font-family: Helvetica;
-  width: 100%;
+  width: ${({ block }: ButtonBaseProps) => block ? '100%' : 'auto'};
   border: none;
   text-shadow: none;
   background: transparent;
   color: orange;
   font-size: 18px;
+  cursor: pointer;
 `;