+++ src/components/Title/index.tsx
@@ -1,16 +1,24 @@
 import styled from "@emotion/styled";
 
-export interface TitleProps {
+export interface TitleProps extends TitleBaseProps {
   label: string;
 }
 
-export function Title({ label }: TitleProps) {
-  return <StyledTitle data-testid="title">{label}</StyledTitle>;
+interface TitleBaseProps {
+  inverted?: boolean
+}
+
+export function Title({ label, inverted }: TitleProps) {
+  return (
+    <StyledTitle data-testid="title" inverted={inverted}>
+      {label}
+    </StyledTitle>
+  )
 }
 
 const StyledTitle = styled.h1`
   font-family: Helvetica;
-  color: white;
+  color: ${({ inverted }: TitleBaseProps) => inverted ? '#333333' : '#FAFAFA'};
   padding: 5px;
   text-align: center;
 `;